requirements.txt
aio-pika==9.4.3
aiohttp==3.9.5
aiormq==6.8.0
aiosignal==1.3.1
alembic==1.13.2
amocrm-api==2.6.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==5.0.1
asyncpg==0.29.0
attrs==23.2.0
black==24.8.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dependency-injector==4.46.0
dnspython==2.6.1
elastic-transport==8.15.0
elasticsearch==8.15.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.10.5
packaging==24.1
pamqp==3.3.0
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==0.22.0
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4
